backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-authtoken>=1.2
django-cors-headers>=4.0
pytz
